Fundamentals game
fist to five.. 3 because I was able to learn html and css on my own. I
did not do well with the javascript. I need to practice all the basics
one by one for more than a day for it to stick. I will be learning
javascript over the weekend at my own pace. I am able to read some js
code, but I am not yet confidant nor comfortable with writing the code
for the instructions given in fundamentals.